+++ src/contents/component/status.tsx
@@ -1,4 +1,4 @@
-import { parseVideo } from "~parse"
+import { sendToBackground } from "@plasmohq/messaging"
 
 export const Parser = () => {
   const vid = document.querySelector("video")
@@ -17,4 +17,33 @@ export const NotParsing = () => {
   return (
     <div className="bg-red-500 absolute top-2 left-2 size-2 rounded-full"></div>
   )
+}
+
+const parseVideo = async () => {
+  const video = document.querySelector("video")
+  if (!video) return
+
+  const titleBox = document.querySelector(
+    "#root > div:nth-child(2) > div > div:nth-child(2) > div"
+  )
+
+  const title = titleBox?.querySelector("a")?.textContent
+  const episode = titleBox?.querySelector(
+    "div div div:first-child"
+  )?.textContent
+
+  const url = window.location.href
+  const speed = video.playbackRate
+  const time = video.currentTime
+  const isPaused = video.paused
+
+  console.log("parsing", { title, episode, url, speed, time, isPaused })
+
+  const res = await sendToBackground({
+    name: "video",
+    body: {
+      msg: "update",
+      data: { title, episode, url, speed, time, isPaused }
+    }
+  })
 }
